Move storyWrapper to a helpers folder
