+++ angular/projects/spark-angular/src/lib/components/sprk-toggle/sprk-toggle.stories.ts
@@ -1,5 +1,5 @@
 import { RouterModule } from '@angular/router';
-import { storyWrapper } from '../../../../../../.storybook/helpers';
+import { storyWrapper } from '../../../../../../.storybook/helpers/storyWrapper';
 import { APP_BASE_HREF } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { SprkToggleModule } from './sprk-toggle.module';
